# src/GestMantIA.API/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "GestMantIA API",
    "description": "API para el sistema de Gestión de Mantenimiento Inteligente Asistido.",
    "termsOfService": "https://example.com/terms",
    "contact": {
      "name": "Soporte GestMantIA",
      "url": "https://gestmantia.com/contact",
      "email": "soporte@gestmantia.com"
    },
    "license": {
      "name": "Licencia de Uso",
      "url": "https://example.com/license"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/TwoFactorAuth/setup": {
      "get": {
        "tags": [
          "TwoFactorAuth"
        ],
        "summary": "Genera la configuración inicial para la autenticación de dos factores.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TwoFactorSetupResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TwoFactorSetupResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TwoFactorSetupResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/TwoFactorAuth/enable": {
      "post": {
        "tags": [
          "TwoFactorAuth"
        ],
        "summary": "Habilita la autenticación de dos factores para el usuario actual.",
        "requestBody": {
          "description": "Solicitud con el código de verificación.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyTwoFactorRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyTwoFactorRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyTwoFactorRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OperationResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/TwoFactorAuth/disable": {
      "post": {
        "tags": [
          "TwoFactorAuth"
        ],
        "summary": "Deshabilita la autenticación de dos factores para el usuario actual.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OperationResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/TwoFactorAuth/verify": {
      "post": {
        "tags": [
          "TwoFactorAuth"
        ],
        "summary": "Verifica un código de autenticación de dos factores.",
        "requestBody": {
          "description": "Solicitud con el código de verificación.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyTwoFactorRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyTwoFactorRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyTwoFactorRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyTwoFactorResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyTwoFactorResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyTwoFactorResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "OperationResult": {
        "type": "object",
        "properties": {
          "Succeeded": {
            "type": "boolean"
          },
          "Success": {
            "type": "boolean",
            "readOnly": true
          },
          "Message": {
            "type": "string",
            "nullable": true
          },
          "Errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "TwoFactorSetupResult": {
        "type": "object",
        "properties": {
          "Succeeded": {
            "type": "boolean"
          },
          "Success": {
            "type": "boolean",
            "readOnly": true
          },
          "Message": {
            "type": "string",
            "nullable": true
          },
          "Errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true,
            "readOnly": true
          },
          "SharedKey": {
            "type": "string",
            "nullable": true
          },
          "QrCodeImageUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "VerifyTwoFactorRequest": {
        "required": [
          "Code"
        ],
        "type": "object",
        "properties": {
          "Code": {
            "maxLength": 10,
            "minLength": 6,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "VerifyTwoFactorResult": {
        "type": "object",
        "properties": {
          "IsValid": {
            "type": "boolean"
          },
          "Message": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Encabezado de autorización JWT usando el esquema Bearer. Ejemplo: \"Authorization: Bearer {token}\"",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}